Add support for serving https-over-tailscale
Setting TAILSCALE_MODE to http (the default) will continue to
serve traffic plain text on port 80. This is useful when you're
serving internal services without externally accessible DNS
entries.

Setting TAILSCALE_MODE to https will enable the new mode where
Centauri behaves in the same way as if it were listening on
TCP: all HTTP requests on port 80 are redirected to HTTPS, and
HTTPS requests on port 443 are served with a certificate from
the certificate provider. This is useful if you want to stop
browsers complaining and have externally accessible DNS entries
for the hosts.

Thanks to @Greboid for the original PR.