Fix wrong routes for http/2 coalesced connections
HTTP/2 allows clients to reuse connections if another host they
want to talk to is covered by the same SSL certificate. Centauri
previously used the SNI field from the TLS handshake to route
requests, which means they always went to the backend the
connection was opened for even if subsequence requests on the
same connection had a different host header. Now requests are
always routed based on the Host header.

Closes #89