Use a copy-on-write map for routes.
The routes and domain fields in the proxy manager were
guarded by a RWMutex, but this lead to an occasional
issue when updating certificates concurrently. The
certificate updater would hold a read lock for the
(potentially long) duration. This is not normally
a problem, unless another goroutine is updating routes
and tries to acquire a write lock. This goroutine
will be blocked until the read finishes (which is
fine), but also all future reads will be blocked
until the write is unblocked and finished.

This blocking of reads caused Centauri to not
serve requests, as it couldn't look up where to
route them.

A copy-on-write map fixes this, and works well
as we typically do a small number of writes and
a large number of reads.